@@ -29,6 +29,7 @@ class CollectionCard extends StatelessWidget {
   final bool defaultView;
   final bool showWishList;
   final bool showSellList;
+  final bool showCompostList;
 
   CollectionCard({
     @required this.connectionLibrary,
@@ -39,237 +40,254 @@ class CollectionCard extends StatelessWidget {
     @required this.defaultView,
     this.showWishList = false,
     this.showSellList = false,
+    this.showCompostList = false,
   });
 
   @override
   Widget build(BuildContext context) {
+    //easy reference
+    AppData provAppDataFalse = Provider.of<AppData>(context, listen: false);
+    //easy scale
+    double width = MediaQuery.of(context).size.width;
     //use the appropriate plant source
     List<PlantData> fullList = (connectionLibrary == false)
-        ? Provider.of<AppData>(context).currentUserPlants
-        : Provider.of<AppData>(context).connectionPlants;
+        ? provAppDataFalse.currentUserPlants
+        : provAppDataFalse.connectionPlants;
 
     //get plants for the collection from the full list
     List<PlantData> collectionPlants = AppData.getPlantsFromList(
         collectionPlantIDs: collection.plants, plants: fullList);
 
-    //now sort these alphabetically if the user has chosen that option
-    if (sortPlants == true) {
-      collectionPlants.sort((a, b) => a.name.compareTo(b.name));
-    }
+    //issue where empty orphan list shows up
+    bool emptyOrphan = (collection.id == DBDefaultDocument.orphaned &&
+        collectionPlants != null &&
+        collectionPlants.length <= 0);
 
-    //get plants for the collection from the full list
-    String numberedList = UIBuilders.shareList(shelfPlants: collectionPlants);
-
-    //note collection plant total is calculated from this list instead of collection.plants
-    //to prevent range errors if an entry is in collection.plants but the plant is deleted
-    int collectionPlantTotal = collectionPlants.length;
-
-    //item count for gridview
-    int itemCountGridView;
-    int itemCountIndexSwap;
-    if (connectionLibrary == true) {
-      print('Connection Libray');
-      //never show the add button in connection view
-      itemCountGridView = collectionPlantTotal;
-      itemCountIndexSwap = itemCountGridView - 1;
-    } else if (DBDefaultDocument.collectionPreventMoveInto
-        .contains(collection.id)) {
-      //my library don't show add button for certain shelves
-      itemCountGridView = collectionPlantTotal;
-      itemCountIndexSwap = itemCountGridView - 1;
-    } else {
-      //otherwise show the add button
-      itemCountGridView = collectionPlantTotal + 1;
-      itemCountIndexSwap = collectionPlantTotal - 1;
-    }
-
-    //convert card color
-    Color colorTheme = convertColor(storedColor: collection.color);
-
-    //plant tile possible parents
-    List<CollectionData> plantTilePossibleParents = (connectionLibrary == false)
-        ? Provider.of<AppData>(context).currentUserCollections
-        : Provider.of<AppData>(context).connectionCollections;
-    //remove the auto generated import collections and hidden
-    List<CollectionData> reducedParents = [];
-    for (CollectionData collection in plantTilePossibleParents) {
-      if (collection.id == DBDefaultDocument.wishList) {
-        if (showWishList == true) reducedParents.add(collection);
-      } else if (collection.id == DBDefaultDocument.sellList) {
-        if (showSellList == true) reducedParents.add(collection);
-      } else if (!DBDefaultDocument.collectionPreventMoveInto
-          .contains(collection.id)) {
-        reducedParents.add(collection);
+    //now if this list is empty, don't display the collection for orphaned
+    if (emptyOrphan == false) {
+      //now sort these alphabetically if the user has chosen that option
+      if (sortPlants == true) {
+        collectionPlants.sort((a, b) => a.name.compareTo(b.name));
       }
-    }
 
-    //*****SET WIDGET VISIBILITY START*****//
+      //get plants for the collection from the full list
+      String numberedList = UIBuilders.shareList(
+          shelfPlants: collectionPlants, shelfName: collection.name);
 
-    //enable dialogs only if library belongs to the current user
-    //don't show for autogenerated cards
-    bool showDeleteButton = (collectionPlantTotal == 0 &&
-        connectionLibrary == false &&
-        !DBDefaultDocument.collectionAutoGen.contains(collection.id));
+      //note collection plant total is calculated from this list instead of collection.plants
+      //to prevent range errors if an entry is in collection.plants but the plant is deleted
+      int collectionPlantTotal = collectionPlants.length;
 
-    //prevent empty gridview build
-    bool hideGridView =
-        ((connectionLibrary == true && itemCountGridView == 0) ||
-            (connectionLibrary == false &&
-                defaultView == true &&
-                itemCountGridView == 0));
+      //item count for gridview
+      int itemCountGridView;
+      int itemCountIndexSwap;
+      if (connectionLibrary == true) {
+        print('Connection Library');
+        //never show the add button in connection view
+        itemCountGridView = collectionPlantTotal;
+        itemCountIndexSwap = itemCountGridView - 1;
+      } else if (DBDefaultDocument.collectionPreventMoveInto
+          .contains(collection.id)) {
+        //my library don't show add button for certain shelves
+        itemCountGridView = collectionPlantTotal;
+        itemCountIndexSwap = itemCountGridView - 1;
+      } else {
+        //otherwise show the add button
+        itemCountGridView = collectionPlantTotal + 1;
+        itemCountIndexSwap = collectionPlantTotal - 1;
+      }
 
-    //*****SET WIDGET VISIBILITY END*****//
+      //convert card color
+      Color colorTheme = convertColor(storedColor: collection.color);
 
-    return TileWhite(
-      child: Consumer<UserData>(builder: (context, user, _) {
-        if (user == null) {
-          return SizedBox();
-        } else {
-          return ExpandableNotifier(
-            //user settings to determine if collection is collapsed by default
-            initialExpanded: user.expandCollection,
-            child: Expandable(
-              expanded: Column(
-                children: <Widget>[
-                  CollectionHeader(
-                    defaultView: defaultView,
-                    connectionLibrary: connectionLibrary,
-                    collection: collection,
-                    colorTheme: colorTheme,
-                    expandedIcon: Icons.keyboard_arrow_up,
-                    numberedList: numberedList,
-                  ),
-                  //provide a delete button if the collection is empty
-                  (showDeleteButton == true)
-                      ? Column(
-                          children: <Widget>[
-                            Provider.of<AppData>(context).showTips == true
-                                ? InfoTip(
-                                    onPress: () {},
-                                    text:
-                                        '${GlobalStrings.collections} display your ${GlobalStrings.plants}.  '
-                                        'They can only be deleted when empty, via the trash button below.  \n\n'
-                                        'Hold down on the ${GlobalStrings.collection} name above to rename it.  \n\n'
-                                        'Tap the arrow button to the right of the name to collapse or expand.  '
-                                        'Hold down on the arrow button to set the colour.  \n\n'
-                                        'You can add a ${GlobalStrings.plant} with the green "+" button below.  '
-                                        'Hold down on a ${GlobalStrings.plant} tile to move it to another Shelf.  '
-                                        'Then tap the ${GlobalStrings.plant} to visit it\'s profile and add more information.  \n\n')
-                                : SizedBox(),
-                            Container(
-                              decoration: kButtonBoxDecoration,
-                              width: double.infinity,
-                              child: FlatButton(
-                                padding: EdgeInsets.all(10.0),
-                                child: CircleAvatar(
-                                  foregroundColor: kGreenDark,
-                                  backgroundColor: Colors.white,
-                                  radius: AppTextSize.medium *
-                                      MediaQuery.of(context).size.width,
-                                  child: Icon(
-                                    Icons.delete_forever,
-                                    size: AppTextSize.huge *
-                                        MediaQuery.of(context).size.width,
-                                  ),
-                                ),
-                                onPressed: () {
-                                  showDialog(
-                                    context: context,
-                                    builder: (BuildContext context) {
-                                      return DialogConfirm(
-                                        title:
-                                            'Remove ${GlobalStrings.collection}',
-                                        text:
-                                            'Are you sure you want to delete this ${GlobalStrings.collection}?',
-                                        buttonText: 'Remove',
-                                        onPressed: () {
-                                          //delete the collection
-                                          CloudDB.deleteDocumentL2(
-                                              collectionL1: DBFolder.users,
-                                              documentL1:
-                                                  Provider.of<AppData>(context)
-                                                      .currentUserInfo
-                                                      .id,
-                                              collectionL2:
-                                                  DBFolder.collections,
-                                              documentL2: collection.id);
-                                          //pop context
-                                          Navigator.pop(context);
-                                        },
-                                      );
-                                    },
-                                  );
-                                },
-                              ),
-                            ),
-                          ],
-                        )
-                      : const SizedBox(),
-                  Builder(
-                    builder: (context) {
-                      //if friend library with no plants, this prevents empty white space
-                      if (hideGridView == true) {
-                        return SizedBox();
-                      } else {
-                        return GridView.builder(
-                          shrinkWrap: true,
-                          //allows scrolling
-                          primary: false,
-                          padding: EdgeInsets.only(bottom: 10.0),
-                          scrollDirection: Axis.vertical,
-                          //add additional button only for collection owner
-                          //no add button for auto generated
-                          itemCount: itemCountGridView,
-                          gridDelegate:
-                              SliverGridDelegateWithFixedCrossAxisCount(
-                                  crossAxisCount: 3),
-                          itemBuilder: (BuildContext context, int index) {
-                            Widget tile;
-                            //build a tile for each plant in the list, beginning index 0
-                            if (index <= itemCountIndexSwap) {
-                              tile = Padding(
-                                padding: EdgeInsets.all(1.0 *
-                                    MediaQuery.of(context).size.width *
-                                    kScaleFactor),
-                                child: PlantTile(
-                                  connectionLibrary: connectionLibrary,
-                                  possibleParents: reducedParents,
-                                  plant: collectionPlants[index],
-                                  collectionID: collection.id,
-                                  communityView: false,
-                                ),
-                              );
-                              //for the last item put an add button
-                            } else {
-                              tile = Padding(
-                                padding: EdgeInsets.all(1.0 *
-                                    MediaQuery.of(context).size.width *
-                                    kScaleFactor),
-                                child: AddPlant(collectionID: collection.id),
-                              );
-                            }
-                            return tile;
-                          },
-                        );
-                      }
-                    },
-                  ),
-                  SizedBox(height: 10.0),
-                ],
-              ),
-              collapsed: CollectionHeader(
-                defaultView: defaultView,
-                connectionLibrary: connectionLibrary,
-                collection: collection,
-                colorTheme: colorTheme,
-                expandedIcon: Icons.keyboard_arrow_down,
-                numberedList: numberedList,
-              ),
-            ),
-          );
+      //plant tile possible parents
+      List<CollectionData> plantTilePossibleParents =
+          (connectionLibrary == false)
+              ? provAppDataFalse.currentUserCollections
+              : provAppDataFalse.connectionCollections;
+      //remove certain auto-generated and hidden to prevent move-into
+      List<CollectionData> reducedParents = [];
+      for (CollectionData collection in plantTilePossibleParents) {
+        if (collection.id == DBDefaultDocument.wishList) {
+          if (showWishList == true) reducedParents.add(collection);
+        } else if (collection.id == DBDefaultDocument.sellList) {
+          if (showSellList == true) reducedParents.add(collection);
+        } else if (collection.id == DBDefaultDocument.compostList) {
+          if (showCompostList == true) reducedParents.add(collection);
+        } else if (!DBDefaultDocument.collectionPreventMoveInto
+            .contains(collection.id)) {
+          reducedParents.add(collection);
         }
-      }),
-    );
+      }
+
+      //*****SET WIDGET VISIBILITY START*****//
+
+      //enable dialogs only if library belongs to the current user
+      //don't show for autogenerated cards
+      bool showDeleteButton = (collectionPlantTotal == 0 &&
+          connectionLibrary == false &&
+          !DBDefaultDocument.collectionAutoGen.contains(collection.id));
+
+      //prevent empty gridview build
+      bool hideGridView =
+          ((connectionLibrary == true && itemCountGridView == 0) ||
+              (connectionLibrary == false &&
+                  defaultView == true &&
+                  itemCountGridView == 0));
+
+      //*****SET WIDGET VISIBILITY END*****//
+
+      return TileWhite(
+        child: Consumer<UserData>(builder: (context, user, _) {
+          if (user == null) {
+            return SizedBox();
+          } else {
+            return ExpandableNotifier(
+              //user settings to determine if collection is collapsed by default
+              initialExpanded: user.expandCollection,
+              child: Expandable(
+                expanded: Column(
+                  children: <Widget>[
+                    CollectionHeader(
+                      defaultView: defaultView,
+                      connectionLibrary: connectionLibrary,
+                      collection: collection,
+                      colorTheme: colorTheme,
+                      expandedIcon: Icons.keyboard_arrow_up,
+                      numberedList: numberedList,
+                    ),
+                    //provide a delete button if the collection is empty
+                    (showDeleteButton == true)
+                        ? Column(
+                            children: <Widget>[
+                              provAppDataFalse.showTips == true
+                                  ? InfoTip(
+                                      onPress: () {},
+                                      text:
+                                          '${GlobalStrings.collections} display your ${GlobalStrings.plants}.  '
+                                          'They can only be deleted when empty, via the trash button below.  \n\n'
+                                          'Hold down on the ${GlobalStrings.collection} name above to rename it.  \n\n'
+                                          'Tap the arrow button to the right of the name to collapse or expand.  '
+                                          'Hold down on the arrow button to set the colour.  \n\n'
+                                          'You can add a ${GlobalStrings.plant} with the green "+" button below.  '
+                                          'Hold down on a ${GlobalStrings.plant} tile to move it to another Shelf.  '
+                                          'Then tap the ${GlobalStrings.plant} to visit it\'s profile and add more information.  \n\n')
+                                  : SizedBox(),
+                              Container(
+                                decoration: kButtonBoxDecoration,
+                                width: double.infinity,
+                                margin: EdgeInsets.all(1.0),
+                                child: FlatButton(
+                                  padding: EdgeInsets.all(10.0),
+                                  child: CircleAvatar(
+                                    foregroundColor: kGreenDark,
+                                    backgroundColor: Colors.white,
+                                    radius: AppTextSize.medium * width,
+                                    child: Icon(
+                                      Icons.delete_forever,
+                                      size: AppTextSize.huge * width,
+                                    ),
+                                  ),
+                                  onPressed: () {
+                                    showDialog(
+                                      context: context,
+                                      builder: (BuildContext context) {
+                                        return DialogConfirm(
+                                          title:
+                                              'Remove ${GlobalStrings.collection}',
+                                          text:
+                                              'Are you sure you want to delete this ${GlobalStrings.collection}?',
+                                          buttonText: 'Remove',
+                                          onPressed: () {
+                                            //delete the collection
+                                            CloudDB.deleteDocumentL2(
+                                                collectionL1: DBFolder.users,
+                                                documentL1:
+                                                    Provider.of<AppData>(
+                                                            context,
+                                                            listen: false)
+                                                        .currentUserInfo
+                                                        .id,
+                                                collectionL2:
+                                                    DBFolder.collections,
+                                                documentL2: collection.id);
+                                            //pop context
+                                            Navigator.pop(context);
+                                          },
+                                        );
+                                      },
+                                    );
+                                  },
+                                ),
+                              ),
+                            ],
+                          )
+                        : const SizedBox(),
+                    Builder(
+                      builder: (context) {
+                        //if friend library with no plants, this prevents empty white space
+                        if (hideGridView == true) {
+                          return SizedBox();
+                        } else {
+                          return GridView.builder(
+                            shrinkWrap: true,
+                            //allows scrolling
+                            primary: false,
+                            padding: EdgeInsets.only(bottom: 5.0),
+                            scrollDirection: Axis.vertical,
+                            //add additional button only for collection owner
+                            //no add button for auto generated
+                            itemCount: itemCountGridView,
+                            gridDelegate:
+                                SliverGridDelegateWithFixedCrossAxisCount(
+                                    crossAxisCount: 3),
+                            itemBuilder: (BuildContext context, int index) {
+                              Widget tile;
+                              //build a tile for each plant in the list, beginning index 0
+                              if (index <= itemCountIndexSwap) {
+                                tile = Container(
+                                  padding: EdgeInsets.all(0.5),
+                                  child: PlantTile(
+                                    connectionLibrary: connectionLibrary,
+                                    possibleParents: reducedParents,
+                                    plant: collectionPlants[index],
+                                    collectionID: collection.id,
+                                    communityView: false,
+                                  ),
+                                );
+                                //for the last item put an add button
+                              } else {
+                                tile = Padding(
+                                  padding: EdgeInsets.all(
+                                      1.0 * width * kScaleFactor),
+                                  child: AddPlant(collectionID: collection.id),
+                                );
+                              }
+                              return tile;
+                            },
+                          );
+                        }
+                      },
+                    ),
+                    SizedBox(height: 10.0),
+                  ],
+                ),
+                collapsed: CollectionHeader(
+                  defaultView: defaultView,
+                  connectionLibrary: connectionLibrary,
+                  collection: collection,
+                  colorTheme: colorTheme,
+                  expandedIcon: Icons.keyboard_arrow_down,
+                  numberedList: numberedList,
+                ),
+              ),
+            );
+          }
+        }),
+      );
+    } else {
+      return SizedBox();
+    }
   }
 }
 
@@ -291,6 +309,10 @@ class CollectionHeader extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    //easy reference
+    AppData provAppDataFalse = Provider.of<AppData>(context, listen: false);
+    //easy scale
+    double width = MediaQuery.of(context).size.width;
     //set plant number
     String collectionPlantTotal = collection.plants.length.toString();
 
@@ -337,10 +359,9 @@ class CollectionHeader extends StatelessWidget {
                               //create data pair map
                               Map data = AppData.updatePairFull(
                                   key: CollectionKeys.name,
-                                  value: Provider.of<AppData>(context)
-                                      .newDataInput);
+                                  value: provAppDataFalse.newDataInput);
                               //upload update to db
-                              Provider.of<CloudDB>(context)
+                              Provider.of<CloudDB>(context, listen: false)
                                   .updateDocumentInCollection(
                                       data: data,
                                       collection: DBFolder.collections,
@@ -349,8 +370,7 @@ class CollectionHeader extends StatelessWidget {
                               Navigator.pop(context);
                             },
                             onChange: (input) {
-                              Provider.of<AppData>(context).newDataInput =
-                                  input;
+                              provAppDataFalse.newDataInput = input;
                             },
                             cancelText: 'Cancel',
                             hintText: collection.name);
@@ -368,22 +388,20 @@ class CollectionHeader extends StatelessWidget {
                         : Center(
                             child: Icon(
                               Icons.star,
-                              size: AppTextSize.large *
-                                  MediaQuery.of(context).size.width,
+                              size: AppTextSize.large * width,
                               color: kGreenDark,
                             ),
                           ),
                   ),
                   Expanded(
                     child: Container(
-                      width: MediaQuery.of(context).size.width * 0.7,
+                      width: width * 0.7,
                       child: Text(
                         collection.name.toUpperCase(),
                         overflow: TextOverflow.fade,
                         textAlign: TextAlign.center,
                         style: TextStyle(
-                          fontSize: AppTextSize.large *
-                              MediaQuery.of(context).size.width,
+                          fontSize: AppTextSize.large * width,
                           fontWeight: AppTextWeight.medium,
                         ),
                       ),
@@ -410,15 +428,11 @@ class CollectionHeader extends StatelessWidget {
                     },
                     child: ExpandableButton(
                       child: CircleAvatar(
-                        radius: 16.0 *
-                            MediaQuery.of(context).size.width *
-                            kScaleFactor,
+                        radius: 16.0 * width * kScaleFactor,
                         backgroundColor: colorTheme,
                         child: Icon(
                           expandedIcon,
-                          size: 30.0 *
-                              MediaQuery.of(context).size.width *
-                              kScaleFactor,
+                          size: 30.0 * width * kScaleFactor,
                           color: AppTextColor.white,
                         ),
                       ),
@@ -430,7 +444,7 @@ class CollectionHeader extends StatelessWidget {
           ),
           Container(
             height: 2.0,
-            width: MediaQuery.of(context).size.width * 0.65,
+            width: width * 0.65,
             color: colorTheme,
           ),
           SizedBox(height: 5.0),
@@ -446,7 +460,7 @@ class CollectionHeader extends StatelessWidget {
               children: <Widget>[
                 Icon(
                   Icons.view_list,
-                  size: AppTextSize.small * MediaQuery.of(context).size.width,
+                  size: AppTextSize.small * width,
                   color: AppTextColor.light,
                 ),
                 SizedBox(
@@ -458,8 +472,7 @@ class CollectionHeader extends StatelessWidget {
                       : '$collectionPlantTotal ${GlobalStrings.plants} on ${GlobalStrings.collection}',
                   style: TextStyle(
                       color: AppTextColor.light,
-                      fontSize: AppTextSize.small *
-                          MediaQuery.of(context).size.width),
+                      fontSize: AppTextSize.small * width),
                 ),
               ],
             ),
