@@ -1,6 +1,5 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:plant_collector/formats/colors.dart';
 import 'package:plant_collector/models/user.dart';
 import 'package:provider/provider.dart';
 import 'package:plant_collector/screens/login/login.dart';
@@ -11,18 +10,18 @@ class RouteScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     //check to see if the current user is still signed in
-    return FutureBuilder<FirebaseUser>(
-        future: Provider.of<UserAuth>(context).getCurrentUser(),
+    return FutureBuilder<User>(
+        future: Provider.of<UserAuth>(context, listen: false).getCurrentUser(),
         builder: (context, snap) {
           if (snap != null) {
-            FirebaseUser user = snap.data;
+            User user = snap.data;
             if (user == null) {
               //if not signed in go to login
               return LoginScreen();
             }
 
             //otherwise save user data and head to main Library screen
-            Provider.of<UserAuth>(context).signedInUser = user;
+            Provider.of<UserAuth>(context, listen: false).signedInUser = user;
             return LibraryScreen(
               userID: user.uid,
               connectionLibrary: false,
@@ -30,7 +29,7 @@ class RouteScreen extends StatelessWidget {
           } else {
             //loading indicator while waiting
             return Scaffold(
-              backgroundColor: kGreenLight,
+              // backgroundColor: kGreenLight,
               body: Center(
                 child: CircularProgressIndicator(),
               ),
