@@ -16,10 +16,10 @@ class AddPhoto extends StatelessWidget {
   Widget build(BuildContext context) {
     double widgetScale = largeWidget ? 1.0 : 0.3;
     return Container(
-      margin: EdgeInsets.only(bottom: largeWidget ? 10.0 : 0.0),
-      padding: EdgeInsets.all(largeWidget ? 4.0 : 0.0),
+      width: double.infinity,
+      margin: EdgeInsets.symmetric(horizontal: largeWidget ? 2.0 : 0.0),
       child: Container(
-        width: MediaQuery.of(context).size.width * 0.94,
+        // width: MediaQuery.of(context).size.width * 0.94,
         decoration: largeWidget
             ? BoxDecoration(
                 gradient: kGradientGreenVerticalDarkMed,
