@@ -19,36 +19,43 @@ class ButtonAdd extends StatelessWidget {
   Widget build(BuildContext context) {
     return GestureDetector(
       child: Container(
-        margin: EdgeInsets.all(
-          5.0 * MediaQuery.of(context).size.width * kScaleFactor,
+        margin: EdgeInsets.only(
+          top: 5.0 * MediaQuery.of(context).size.width * kScaleFactor,
         ),
         padding: EdgeInsets.symmetric(vertical: 10.0),
         decoration: BoxDecoration(
             boxShadow: kShadowBox,
             borderRadius: BorderRadius.all(
               Radius.circular(
-                5.0,
+                10.0,
               ),
             ),
             gradient: kGradientGreenVerticalDarkMed),
         child: Row(
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
-            Icon(
-              icon != null ? icon : Icons.add,
-              size:
-                  scale * AppTextSize.huge * MediaQuery.of(context).size.width,
-              color: textColor != null ? textColor : AppTextColor.white,
-            ),
             Text(
               '   $buttonText',
+              textAlign: TextAlign.center,
               style: TextStyle(
                 color: textColor != null ? textColor : AppTextColor.white,
                 fontSize: scale *
                     AppTextSize.huge *
                     MediaQuery.of(context).size.width,
-                fontWeight: AppTextWeight.medium,
-//                shadows: kShadowText,
+                fontWeight: AppTextWeight.heavy,
+              ),
+            ),
+            Align(
+              alignment: Alignment.centerLeft,
+              child: Container(
+                padding: EdgeInsets.symmetric(horizontal: 5),
+                child: Icon(
+                  icon != null ? icon : Icons.add_circle,
+                  size: scale *
+                      AppTextSize.huge *
+                      MediaQuery.of(context).size.width,
+                  color: textColor != null ? textColor : AppTextColor.white,
+                ),
               ),
             ),
           ],
