@@ -5,9 +5,11 @@ import 'package:plant_collector/widgets/tile_white.dart';
 class SectionHeader extends StatelessWidget {
   final String title;
   final Widget leading;
+  final Widget trailing;
   SectionHeader({
     @required this.title,
     this.leading,
+    this.trailing = const SizedBox(),
   });
 
   @override
@@ -20,13 +22,18 @@ class SectionHeader extends StatelessWidget {
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
             (leading == null) ? SizedBox() : leading,
-            Text(
-              title != null ? title : '',
-              style: TextStyle(
-                fontSize: AppTextSize.huge * MediaQuery.of(context).size.width,
-                fontWeight: AppTextWeight.medium,
+            Expanded(
+              child: Text(
+                title != null ? title : '',
+                textAlign: TextAlign.center,
+                style: TextStyle(
+                  fontSize:
+                      AppTextSize.huge * MediaQuery.of(context).size.width,
+                  fontWeight: AppTextWeight.medium,
+                ),
               ),
             ),
+            trailing,
           ],
         ),
       ),
